+++ components/role-os/dashboards/index.ts
@@ -1,0 +1,1 @@
+export { default as RoleDashboardExperience } from "./RoleDashboardExperience";